requirements.txt
streamlit>=1.32.0
yt-dlp>=2024.1.1
google-api-python-client>=2.118.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
Pillow>=10.0.0
numpy>=1.24.0
supabase>=2.0.0
